requirements.txt
asgiref==3.2.10
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.9
django-filter==2.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
Markdown==3.2.2
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
psycopg2=2.8.5

psycopg2~=2.8.5